Remove white space above/under page when scrolling
This change was made to help with aesthetics of the page. This issue occurred because the background color of the body was set to white (default). Therefore, when overscrolling the page, there would be white space above or under the page. This was fixed by setting a background color to the body element that matches our header and footer. There was also changes made to the call to action container to remove the margins and add it instead to the sign-up container to make this change work.